componente de formulario de curso separado
